split indicator_data_extraction tests into separate files
